Move the znodehandler to the commons/zookeeper package
